+++ src/main/java/com/negocore/application/dto/response/CashRegisterResponseDTO.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @AllArgsConstructor
@@ -18,9 +19,9 @@ public class CashRegisterResponseDTO {
 
     private Long id;
     private Long businessId;
-    private Double openingAmount;
-    private Double expectedAmount;
-    private Double closingAmount;
+    private BigDecimal openingAmount;
+    private BigDecimal expectedAmount;
+    private BigDecimal closingAmount;
     private CashRegisterStatus status;
     private LocalDateTime openingAt;
     private LocalDateTime closingAt;
