+++ src/main/java/com/negocore/application/dto/response/ProductResponseDTO.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @AllArgsConstructor
@@ -20,8 +21,8 @@ public class ProductResponseDTO {
     private Long categoryId;
     private String name;
     private String sku;
-    private Double costPrice;
-    private Double salePrice;
+    private BigDecimal costPrice;
+    private BigDecimal salePrice;
     private Integer stock;
     private Integer minStockAlert;
     private Boolean active;
