+++ src/main/java/com/negocore/domain/api/ICashRegisterServicePort.java
@@ -2,8 +2,10 @@ package com.negocore.domain.api;
 
 import com.negocore.domain.model.CashRegister;
 
+import java.math.BigDecimal;
+
 public interface ICashRegisterServicePort {
 
-    CashRegister openCashRegister(Long businessId, Double openingAmount);
+    CashRegister openCashRegister(Long businessId, BigDecimal openingAmount);
 
 }
