+++ src/main/java/com/negocore/domain/model/Product.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -18,8 +19,8 @@ public class Product {
     private Long categoryId;
     private String name;
     private String sku;
-    private Double costPrice;
-    private Double salePrice;
+    private BigDecimal costPrice;
+    private BigDecimal salePrice;
     private Integer stock;
     private Integer minStockAlert;
     private Boolean active;
